+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 -   The `date_range` argument in `CDR()` now accepts a length one vector and throws an error if a vector longer than 2 elements is passed.
+-   `sea_ice_index()` returns a complete time series, making implicit missing values explicit.
 
 ## Bug Fixes